// SDKs/googleapis.com/networkmanagement/v1beta1/typescript/src/sdk/models/shared/endpoint.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import { AppEngineVersionEndpoint } from "./appengineversionendpoint";
import { CloudFunctionEndpoint } from "./cloudfunctionendpoint";
import { CloudRunRevisionEndpoint } from "./cloudrunrevisionendpoint";


export enum EndpointNetworkTypeEnum {
    NetworkTypeUnspecified = "NETWORK_TYPE_UNSPECIFIED",
    GcpNetwork = "GCP_NETWORK",
    NonGcpNetwork = "NON_GCP_NETWORK"
}


// Endpoint
/** 
 * Source or destination of the Connectivity Test.
**/
export class Endpoint extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=appEngineVersion" })
  appEngineVersion?: AppEngineVersionEndpoint;

  @SpeakeasyMetadata({ data: "json, name=cloudFunction" })
  cloudFunction?: CloudFunctionEndpoint;

  @SpeakeasyMetadata({ data: "json, name=cloudRunRevision" })
  cloudRunRevision?: CloudRunRevisionEndpoint;

  @SpeakeasyMetadata({ data: "json, name=cloudSqlInstance" })
  cloudSqlInstance?: string;

  @SpeakeasyMetadata({ data: "json, name=gkeMasterCluster" })
  gkeMasterCluster?: string;

  @SpeakeasyMetadata({ data: "json, name=instance" })
  instance?: string;

  @SpeakeasyMetadata({ data: "json, name=ipAddress" })
  ipAddress?: string;

  @SpeakeasyMetadata({ data: "json, name=network" })
  network?: string;

  @SpeakeasyMetadata({ data: "json, name=networkType" })
  networkType?: EndpointNetworkTypeEnum;

  @SpeakeasyMetadata({ data: "json, name=port" })
  port?: number;

  @SpeakeasyMetadata({ data: "json, name=projectId" })
  projectId?: string;
}